+++ src/main.rs
@@ -1,3 +1,4 @@
+mod reflinks_query;
 mod server;
 
 use std::path::PathBuf;